events-api/pyproject.toml
[tool.poetry]
name = "events-api"
version = "0.3.0"
description = ""
authors = ["David Escobar <david.escobar@ug.uchile.cl>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "3.10.9"
aiosqlite = "0.19.0"
anyio = "3.6.1"
backports-cached-property = "1.0.2"
beautifulsoup4 = "4.11.1"
certifi = "2022.6.15"
charset-normalizer = "2.1.0"
click = "8.1.3"
databases = "0.7.0"
fastapi = "0.95.2"
graphql-core = "3.2.1"
greenlet = "2.0.2"
h11 = "0.13.0"
httptools = "0.4.0"
idna = "3.3"
lxml = "4.9.1"
numpy = "1.23.1"
orm = "0.3.1"
pandas = "1.4.3"
pydantic = "1.9.1"
pygments = "2.12.0"
python-dateutil = "2.8.2"
python-decouple = "3.6"
python-dotenv = "0.20.0"
python-levenshtein = "0.12.2"
python-multipart = "0.0.5"
pytz = "2022.1"
pyyaml = "6.0"
requests = "2.28.1"
six = "1.16.0"
sniffio = "1.2.0"
soupsieve = "2.3.2.post1"
sqlalchemy = "1.4.48"
starlette = "0.27.0"
strawberry-graphql = "0.177.1"
thefuzz = "0.19.0"
typesystem = "0.3.1"
typing-extensions = "4.3.0"
urllib3 = "1.26.11"
uvicorn = "0.18.2"
uvloop = "0.16.0"
watchfiles = "0.16.1"
websockets = "10.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
